code refinement

+++ src/main/java/dev/masa/masuitewarps/bungee/controllers/DeleteController.java
@@ -1,15 +1,13 @@
 package dev.masa.masuitewarps.bungee.controllers;
 
 import dev.masa.masuitecore.bungee.chat.Formator;
-import dev.masa.masuitecore.core.channels.BungeePluginChannel;
 import dev.masa.masuitecore.core.configuration.BungeeConfiguration;
 import dev.masa.masuitecore.core.models.MaSuitePlayer;
 import dev.masa.masuitewarps.bungee.MaSuiteWarps;
 import dev.masa.masuitewarps.core.models.Warp;
-import net.md_5.bungee.api.config.ServerInfo;
 import net.md_5.bungee.api.connection.ProxiedPlayer;
 
-import java.util.Map;
+import java.util.UUID;
 
 /**
  * @author Masa
@@ -24,33 +22,32 @@ public class DeleteController {
         this.plugin = plugin;
     }
 
-    public void deleteWarp(ProxiedPlayer player, String warpName) {
+    public void deleteWarp(ProxiedPlayer sender, UUID ownerId, String warpName) {
         Warp warp = plugin.getWarpService().getWarp(warpName);
         if (warp == null) {
-            plugin.formator.sendMessage(player, plugin.warpNotFound);
+            plugin.formator.sendMessage(sender, plugin.warpNotFound);
             return;
         }
+
+        if (!warp.getOwner().equals(ownerId)) {
+            formator.sendMessage(sender, config.load("warps", "messages.yml").getString("not-owner"));
+            return;
+        }
+
         if (plugin.getWarpService().removeWarp(warp)) {
-            plugin.formator.sendMessage(player, plugin.warpDeleted.replace("%warp%", warp.getName()));
-            for (Map.Entry<String, ServerInfo> entry : plugin.getProxy().getServers().entrySet()) {
-                ServerInfo serverInfo = entry.getValue();
-                serverInfo.ping((result, error) -> {
-                    if (error == null) {
-                        new BungeePluginChannel(plugin, serverInfo, "DelWarp", warpName).send();
-                    }
-                });
-            }
+            plugin.formator.sendMessage(sender, plugin.warpDeleted.replace("%warp%", warp.getName()));
         } else {
-            plugin.formator.sendMessage(player, "&cAn error occurred. Please check console for more details");
+            plugin.formator.sendMessage(sender, "&cAn error occurred. Please check console for more details");
         }
     }
 
-    public void deleteOthersWarp(ProxiedPlayer player, String warpName, String ownerName) {
+    public void deleteOthersWarp(ProxiedPlayer sender, String ownerName, String warpName) {
         MaSuitePlayer mspOwner = plugin.getApi().getPlayerService().getPlayer(ownerName);
         if (mspOwner == null) {
-            formator.sendMessage(player, config.load("warps", "messages.yml").getString("player-not-found"));
+            formator.sendMessage(sender, config.load("warps", "messages.yml").getString("player-not-found"));
             return;
         }
-        deleteWarp(player, warpName);
+
+        deleteWarp(sender, mspOwner.getUniqueId(), warpName);
     }
 }
